move fn to customHook
