+++ src/hooks/useMessagesChat.ts
@@ -1,11 +1,15 @@
 import { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { OnClickBtnDelteType } from '../components/MessagesList/types';
+import {
+  OnClickBtnDelteType,
+  CheckUserType,
+  IsOwnMsgsType,
+} from '../components/MessagesList/types';
 import { selectMsgsChat } from '../redux/chatSelector';
 import { deleteMsg } from '../redux/chatSlice';
 import { ChatAppData } from '../redux/types';
 
-const useMessagesChat = () => {
+const useMessagesChat = (userLogged: string) => {
   const dispatch = useDispatch();
   const [isFinishTransition, setIsFinishtransition] = useState(false);
 
@@ -15,11 +19,17 @@ const useMessagesChat = () => {
     dispatch(deleteMsg(idToDelete));
   const setTrueFinishTrans = () => setIsFinishtransition(true);
 
+  const checkUser: CheckUserType = (userMsgs) => userMsgs === userLogged;
+  const isOwnMsgs: IsOwnMsgsType = (userMsgs) =>
+    checkUser(userMsgs) ? 'onRight' : 'onLeft';
+
   return {
     isFinishTransition,
     setTrueFinishTrans,
     onClickBtnDelte,
     msgsChat,
+    checkUser,
+    isOwnMsgs,
   };
 };
